Convert select transit agency screen to use UITableViewController
